pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astro-virtual-lab"
version = "1.0.0"
authors = [
    { name = "Ioana Ciuca", email = "iciuca@stanford.edu" },
]
maintainers = [
    { name = "Ioana Ciuca", email = "iciuca@stanford.edu" },
]
description = "A virtual lab of large language model agents for astronomical research."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "notebook>=7.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "tiktoken>=0.6.0",
    "tqdm>=4.66.0",
    "typed-argument-parser>=1.8.0",
    "astropy>=6.0.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "scipy>=1.12.0",
    "PyYAML>=6.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
keywords = [
    "virtual lab",
    "large language model",
    "LLM",
    "scientific research",
    "automated science",
    "astronomy",
    "astrophysics",
    "data analysis"
]

[project.optional-dependencies]
astronomy = [
    "astroquery",
    "photutils",
    "specutils",
    "astroML",
    "sunpy",
    "regions"
]

[tool.hatch.build.targets.wheel]
packages = ["src/virtual_lab"]

[project.urls]
Homepage = "https://github.com/errai34/astro-virtual-lab"
Issues = "https://github.com/errai34/astro-virtual-lab/issues"
